typing + backwardcompat
